Move search tests into a new file.
